Remove docs folder from repo root
I am adding some automation to the generation of documentation so I am
a. moving documentation to `cmd` for each sub package to better support
multiple packages down the line and b. committing the docs to the nox
wiki so that its easier to and more centralized and not part of this git
history.